+++ src/api/apiPostsSlice.ts
@@ -1,6 +1,6 @@
 import { FetchBaseQueryError } from '@reduxjs/toolkit/query';
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { FetchService } from '../helpers';
+import { FetchService } from '../services';
 import {
   FetchResponse,
   FullPostCard,
