+++ src/api/index.ts
@@ -1,2 +1,1 @@
-export * from './LocalStorageService.ts';
-export * from './PostsService.ts';
+export * from './apiPostsSlice';
