+++ src/components/PaginationUI/PaginationUI.tsx
@@ -1,6 +1,6 @@
 import { QueryStatus } from '@reduxjs/toolkit/query';
 import { useContext, useMemo } from 'react';
-import { PostsListSearchQueryContext } from '../../context/PostsListContext.tsx';
+import { SearchQueryContext } from '../../context';
 import { useGetAllPostsQuery } from '../../api';
 import PaginationButtons from '../PaginationButtons/PaginationButtons.tsx';
 import PaginationCountInfo from '../PaginationCountInfo/PaginationCountInfo.tsx';
@@ -8,7 +8,7 @@ import PaginationSelect from '../PaginationSelect/PaginationSelect.tsx';
 import style from './PaginationUI.module.css';
 
 function PaginationUI() {
-  const searchQuery = useContext(PostsListSearchQueryContext);
+  const searchQuery = useContext(SearchQueryContext);
   const { data, status } = useGetAllPostsQuery(searchQuery);
 
   const disabled = useMemo(() => status === QueryStatus.pending, [status]);
