--- src/enums/Statuses.ts
@@ -1,6 +1,0 @@
-export enum LoadingStatuses {
-  Idle = 'idle',
-  Pending = 'pending',
-  Fulfilled = 'fulfilled',
-  Rejected = 'rejected',
-}
