--- src/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from './getURLSearchParams.ts';
-export * from './withQueryParam.tsx';
-export * from './fetchFromApi.ts';
