+++ src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './FetchService';
+export * from './LocalStorageService';
