+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './test-utils';
+export * from './render-with-query-param';
